+++ src/hooks/useIsPoolsPage.ts
